MAKE IT FASTER!

+++ src/Filter.hh
@@ -16,7 +16,7 @@ static unsigned numOfFilts;
 class Filter {
 
     private:
-    float a0,a1,a2,b0,b1,b2;
+    double a0,a1,a2,b0,b1,b2;
     double Q = 0.25;
     //double x1 = 0, x2 = 0, y1 = 0, y2 = 0;
     double s1 = 0, s2 = 0;
